packages/assisted-living/dev-utils/pyproject.toml
[project]
name = "dev-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.1",
    "python-dotenv>=1.1.1",
]
